+++ CHANGELOG.md
@@ -27,8 +27,10 @@
 ## 0.5.3 - unreleased
 - Fix INTERNAL ERROR in `FastLSolve.solveL1Straight()` caused by bug in `DLCP.solve1()`.
   [#133](https://github.com/tzaeschke/ode4j/issues/133)
-- Fix formatting of degub/info/error message.
+- Fix formatting of debug/info/error message.
   [#135](https://github.com/tzaeschke/ode4j/issues/135)
+- QOL: allow calling destroy() multiple times.
+  [#139](https://github.com/tzaeschke/ode4j/pull/139)
 
 ## 0.5.2 - 2023-10-07
 - Fix DVector3.cross() returning always 0. [#128](https://github.com/tzaeschke/ode4j/issues/128)